Packages split into icp-connect-core, icp-connect-react, and icp-connect-react-native
