+++ frontend/apps/webapp/next.config.js
@@ -1,22 +1,23 @@
+const webpack = require("webpack");
+const transpileModules = require('next-transpile-modules');
+
 const DFXWebPackConfig = require("./dfx.webpack.config");
 DFXWebPackConfig.bootstrap("../../..");
 
-const { withExpo } = require("@expo/next-adapter");
-const webpack = require("webpack");
+const withTM = transpileModules(['icp-connect-core', 'icp-connect-react']);
 
 // Make DFX_NETWORK available to Web Browser with default "local" if DFX_NETWORK is undefined
 const EnvPlugin = new webpack.EnvironmentPlugin({
-  DFX_NETWORK: "local",
+    DFX_NETWORK: "local",
 });
 
-module.exports = withExpo({
-  webpack: (config, { buildId, dev, isServer, defaultLoaders, webpack }) => {
-    // Plugin
-    config.plugins.push(EnvPlugin);
+module.exports = withTM({
+    webpack: (config) => {
+        // Plugin
+        config.plugins.push(EnvPlugin);
 
-    return config;
-  },
-  transpilePackages: ["expo-web-browser", "expo-secure-store"],
-  output: "export",
-  distDir: "build",
+        return config;
+    },
+    output: "export",
+    distDir: "build",
 });
