+++ frontend/apps/webapp/package.json
@@ -15,21 +15,20 @@
     "@dfinity/candid": "^0.20.2",
     "@dfinity/identity": "^0.20.2",
     "@dfinity/principal": "^0.20.2",
-    "@expo/next-adapter": "^5.0.2",
     "@radix-ui/react-avatar": "^1.0.4",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-scroll-area": "^1.0.5",
     "@radix-ui/react-slot": "^1.0.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
-    "expo": "^49.0.21",
     "firebase": "^10.7.1",
-    "icp-connect": "*",
+    "icp-connect-core": "*",
+    "icp-connect-react": "*",
     "lucide-react": "^0.294.0",
     "next": "^14.0.4",
+    "next-transpile-modules": "^10.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-native-web": "^0.19.9",
     "tailwind-merge": "^2.1.0",
     "tailwindcss-animate": "^1.0.7"
   },
@@ -40,6 +39,7 @@
     "autoprefixer": "^10.0.1",
     "eslint": "^8",
     "eslint-config-next": "^14.0.4",
+    "next-transpile-modules": "^10.0.1",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
     "typescript": "^5"
