+++ frontend/apps/webapp/tsconfig.json
@@ -22,9 +22,6 @@
       "dsnap/*": [
         "./src/*"
       ],
-      "@/packages/*": [
-        "./src/packages/*"
-      ],
       "@/declarations": [
         "./src/declarations"
       ],
