+++ frontend/packages/icp-connect-core/client/client.ts
