+++ frontend/packages/icp-connect-core/client/client.types.ts
