+++ frontend/packages/icp-connect-core/client/index.ts
