+++ frontend/packages/icp-connect-core/identity-providers/identity-provider.interface.ts
