+++ frontend/packages/icp-connect-core/identity-providers/index.ts
@@ -1,4 +1,3 @@
 export * from "./identity-provider.interface";
 export * from "./native-identity-provider.interface";
 export * from "./internet-identity";
-export * from "./internet-identity-react-native";
