+++ frontend/packages/icp-connect-core/identity-providers/internet-identity/index.ts
