+++ frontend/packages/icp-connect-core/identity-providers/internet-provider.constants.ts
