+++ frontend/packages/icp-connect-core/identity-providers/native-identity-provider.interface.ts
