+++ frontend/packages/icp-connect-core/webpack.config.js
