+++ frontend/packages/icp-connect-react-native/identity-providers/index.ts
@@ -1,0 +1,1 @@
+export * from "./internet-identity-react-native";
