+++ frontend/packages/icp-connect-react-native/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "icp-connect-react-native",
+  "version": "0.1.0",
+  "main": "index.ts",
+  "author": "MichiLabs",
+  "license": "MIT",
+  "private": false,
+  "dependencies": {
+    "@dfinity/agent": "^0.20.2",
+    "@dfinity/auth-client": "^0.20.2",
+    "@dfinity/identity": "^0.20.2",
+    "@dfinity/principal": "^0.20.2",
+    "@react-native-async-storage/async-storage": "^1.21.0",
+    "expo-secure-store": "^12.5.0",
+    "expo-web-browser": "^12.5.0",
+    "icp-connect-core": "*"
+  },
+  "devDependencies": {
+    "ts-loader": "^9.5.1",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
+  }
+}
