+++ frontend/packages/icp-connect-react/context.tsx
@@ -1,6 +1,5 @@
+import { Client } from "icp-connect-core/client";
 import React, { ReactNode, createContext } from "react";
-
-import { Client } from "../core/client";
 
 export type IcpConnectContextType<T extends Record<string, any>> = {
   client: Client<T>;
@@ -11,9 +10,7 @@ export type IcpConnectContextProviderProps<T extends Record<string, any>> = {
   client: Client<T>;
 };
 
-export const IcpConnectContext = createContext<
-  IcpConnectContextType<Record<string, any>>
->({} as any);
+export const IcpConnectContext = createContext<IcpConnectContextType<Record<string, any>>>({} as any);
 
 export const IcpConnectContextProvider = <T extends Record<string, any>>({
   children,
