+++ frontend/packages/icp-connect-react/hooks/index.ts
