+++ frontend/packages/icp-connect-react/hooks/useActor.ts
@@ -1,11 +1,8 @@
+import { IcpConnectContext, IcpConnectContextType } from "../context";
+import { ActorMap } from "icp-connect-core/client";
 import { useContext } from "react";
 
-import { IcpConnectContext, IcpConnectContextType } from "../context";
-import { ActorMap } from "../../core/client";
-
-export const useActor = <T extends Record<string, any>>(
-  name: keyof T
-): ActorMap<T>[keyof T] => {
+export const useActor = <T extends Record<string, any>>(name: keyof T): ActorMap<T>[keyof T] => {
   const { client } = useContext(IcpConnectContext) as IcpConnectContextType<T>;
 
   const actor = client.getActor(name);
