+++ frontend/packages/icp-connect-react/hooks/useAuth.ts
