+++ frontend/packages/icp-connect-react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "icp-connect",
+  "name": "icp-connect-react",
   "version": "0.1.0",
   "main": "index.ts",
   "author": "MichiLabs",
@@ -10,9 +10,7 @@
     "@dfinity/auth-client": "^0.20.2",
     "@dfinity/identity": "^0.20.2",
     "@dfinity/principal": "^0.20.2",
-    "@react-native-async-storage/async-storage": "^1.21.0",
-    "expo-secure-store": "^12.5.0",
-    "expo-web-browser": "^12.5.0",
+    "icp-connect-core": "*",
     "react": "^18.2.0"
   },
   "devDependencies": {
