--- frontend/packages/icp-connect/core/index.ts
@@ -1,1 +1,0 @@
-export * from "./client/client";
